% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchaumOrganData.R
\docType{data}
\name{brain}
\alias{brain}
\title{Differential expression analysis using Schaum et al. (2020) brain data.}
\format{
A DESeq2 object containing results of differential expression analysis:
\describe{
\item{assays}{List of matrices with raw and normalized count data}
\item{rowRanges}{Genomic ranges for each gene}
\item{colData}{Data frame with sample metadata}
\item{design}{Formula representing the experimental design}
\item{dispersions}{Estimated dispersion values}
\item{results}{Data frame with differential expression results}
}
}
\source{
\url{doi:10.1038/s41586-020-2499-y}
}
\usage{
brain
}
\description{
Expression data for mouse brain samples across time
}
\keyword{datasets}
