% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchaumOrganData.R
\docType{data}
\name{kidney_metadata}
\alias{kidney_metadata}
\title{Sample meta data for a longitudinal mouse kidney dataset}
\format{
A data frame  with 30 rows and 3 variables:
\describe{
\item{tissue}{name of organ the sample comes from}
\item{age}{age of the mouse in months when sampled}
\item{sex}{sex of mouse from which sample was taken}
\item{age_sex}{combination of age and sex categorical variable}
\item{tissue_age}{combination of tissue and age categorical variable}
\item{condition}{combination of age, sex, and tissue categorical variable}
\item{sample}{unique identifier for sample from original study}
}
}
\source{
\url{doi:10.1038/s41586-020-2499-y}
}
\usage{
kidney_metadata
}
\description{
Sample meta data for the kidney data
}
\keyword{datasets}
