% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAcrossHeart.R
\name{plotAcrossHeart}
\alias{plotAcrossHeart}
\title{plotAcrossHeart}
\usage{
plotAcrossHeart(x)
}
\arguments{
\item{x}{A string containing a valid Mouse Genome Informatics (MGI) ID that is found in the dataset}
}
\value{
A figure plotting the gene expression of filtered TD-data from the Schaum et al. (2020) heart dataset across samples from both sexes of mice over time
}
\description{
Creates a figure plotting the gene expression using filtered TD-data from the Schaum et al. (2020) heart dataset across samples from both sexes of mice over time
}
