% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchaumOrganData-package.R
\docType{package}
\name{SchaumOrganData-package}
\alias{SchaumOrganData}
\alias{SchaumOrganData-package}
\title{SchaumOrganData: Contains RNA-seq data for Schaum et al. (2020) study on mouse organ gene expression over time}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sayumi York \email{syork@ndm.edu}

}
\keyword{internal}
